+++ angular-workspace/projects/ni/nimble-angular/src/directives/number-field/nimble-number-field.directive.ts
@@ -1,8 +1,10 @@
 import { Directive, ElementRef, Input, Renderer2 } from '@angular/core';
 import type { NumberField } from '@ni/nimble-components/dist/esm/number-field';
+import { NumberFieldAppearance } from '@ni/nimble-components/dist/esm/number-field/types';
 import { BooleanValueOrAttribute, NumberValueOrAttribute, toBooleanProperty, toNumberProperty } from '../utilities/template-value-helpers';
 
 export type { NumberField };
+export { NumberFieldAppearance };
 
 /**
  * Directive to provide Angular integration for the number field.
@@ -11,6 +13,14 @@ export type { NumberField };
     selector: 'nimble-number-field'
 })
 export class NimbleNumberFieldDirective {
+    public get appearance(): NumberFieldAppearance {
+        return this.elementRef.nativeElement.appearance;
+    }
+
+    @Input() public set appearance(value: NumberFieldAppearance) {
+        this.renderer.setProperty(this.elementRef.nativeElement, 'appearance', value);
+    }
+
     public get readOnly(): boolean {
         return this.elementRef.nativeElement.readOnly;
     }