+++ packages/nimble-components/src/number-field/styles.ts
@@ -1,7 +1,6 @@
 import { css } from '@microsoft/fast-element';
 import { display } from '@microsoft/fast-foundation';
 import {
-    borderColor,
     borderRgbPartialColor,
     borderHoverColor,
     borderWidth,
@@ -9,15 +8,17 @@ import {
     bodyDisabledFontColor,
     controlHeight,
     fillSelectedRgbPartialColor,
-    iconSize,
     controlLabelFont,
     controlLabelFontColor,
     labelHeight,
     smallDelay,
     bodyFont,
     failColor,
-    standardPadding
+    standardPadding,
+    controlLabelDisabledFontColor
 } from '../theme-provider/design-tokens';
+import { appearanceBehavior } from '../utilities/style/appearance';
+import { NumberFieldAppearance } from './types';
 
 export const styles = css`
     ${display('inline-block')}
@@ -42,6 +43,10 @@ export const styles = css`
         font: ${controlLabelFont};
     }
 
+    :host([disabled]) .label {
+        color: ${controlLabelDisabledFontColor};
+    }
+
     .root {
         box-sizing: border-box;
         position: relative;
@@ -49,7 +54,7 @@ export const styles = css`
         flex-direction: row;
         justify-content: center;
         border-radius: 0px;
-        border-bottom: ${borderWidth} solid rgba(${borderRgbPartialColor}, 0.3);
+        border: 0px solid rgba(${borderRgbPartialColor}, 0.3);
         gap: calc(${standardPadding} / 2);
     }
 
@@ -57,9 +62,8 @@ export const styles = css`
         border-bottom-color: ${borderHoverColor};
     }
 
-    :host([disabled]) .root,
-    :host([disabled]) .root:hover {
-        border-bottom: ${borderWidth} solid ${bodyDisabledFontColor};
+    :host([disabled]) .root {
+        border-color: rgba(${borderRgbPartialColor}, 0.1);
     }
 
     .root::before {
@@ -108,7 +112,7 @@ export const styles = css`
         font: inherit;
         background: transparent;
         color: inherit;
-        height: calc(${controlHeight} - ${borderWidth});
+        height: ${controlHeight};
         width: 100%;
         border: none;
         padding: 0px;
@@ -140,23 +144,59 @@ export const styles = css`
 
     .controls {
         display: flex;
-        flex-direction: column;
-    }
-
-    .step-up,
-    .step-down {
-        display: inline-flex;
-        height: 15px;
-        width: 15px;
-        cursor: pointer;
+        flex-direction: row-reverse;
         justify-content: center;
         align-items: center;
     }
 
-    .step-up svg,
-    .step-down svg {
-        height: ${iconSize};
-        width: ${iconSize};
-        fill: ${borderColor};
+    .step-up-down-button {
+        ${controlHeight.cssCustomProperty}: 24px;
     }
-`;
+`.withBehaviors(
+        appearanceBehavior(
+            NumberFieldAppearance.underline,
+            css`
+            .root {
+                border-bottom-width: ${borderWidth};
+                padding-top: ${borderWidth};
+                padding-left: ${borderWidth};
+                padding-right: ${borderWidth};
+            }
+
+            .control {
+                height: calc(${controlHeight} - 2 * ${borderWidth});
+            }
+        `
+        ),
+        appearanceBehavior(
+            NumberFieldAppearance.block,
+            css`
+            .root {
+                background-color: rgba(${borderRgbPartialColor}, 0.1);
+                padding-left: ${borderWidth};
+                padding-right: ${borderWidth};
+            }
+
+            .root:focus-within,
+            :host(.invalid) .root {
+                border-bottom-width: ${borderWidth};
+            }
+
+            :host([disabled]) .root {
+                background-color: rgba(${borderRgbPartialColor}, 0.07);
+            }
+        `
+        ),
+        appearanceBehavior(
+            NumberFieldAppearance.outline,
+            css`
+            .root {
+                border-width: ${borderWidth};
+            }
+
+            .control {
+                height: calc(${controlHeight} - 2 * ${borderWidth});
+            }
+        `
+        )
+    );