+++ packages/nimble-components/src/number-field/tests/number-field.stories.ts
@@ -3,10 +3,12 @@ import type { Meta, StoryObj } from '@storybook/html';
 import { withXD } from 'storybook-addon-xd-designs';
 import { createUserSelectedThemeStory } from '../../utilities/tests/storybook';
 import '../../all-components';
+import { NumberFieldAppearance } from '../types';
 
 interface NumberFieldArgs {
     label: string;
     value: number;
+    appearance: NumberFieldAppearance;
     disabled: boolean;
 }
 
@@ -32,20 +34,45 @@ const metadata: Meta<NumberFieldArgs> = {
         <nimble-number-field
             placeholder="${x => x.label}"
             value="${x => x.value}"
+            appearance="${x => x.appearance}"
             ?disabled="${x => x.disabled}"
         >
             ${x => x.label}
         </nimble-number-field>
     `),
+    argTypes: {
+        appearance: {
+            options: Object.values(NumberFieldAppearance),
+            control: { type: 'radio' }
+        }
+    },
     args: {
         label: 'default label',
         value: 42,
+        appearance: NumberFieldAppearance.underline,
         disabled: false
     }
 };
 
 export default metadata;
 
-export const numberField: StoryObj<NumberFieldArgs> = {
-    args: { label: 'Number Field' }
+export const underlineNumberField: StoryObj<NumberFieldArgs> = {
+    args: {
+        label: 'Underline Number Field',
+        appearance: NumberFieldAppearance.underline
+    }
+};
+
+export const outlineNumberField: StoryObj<NumberFieldArgs> = {
+    args: {
+        label: 'Outline Number Field',
+        appearance: NumberFieldAppearance.outline
+    }
+};
+
+export const blockNumberField: StoryObj<NumberFieldArgs> = {
+    args: {
+        label: 'Block Number Field',
+        appearance: NumberFieldAppearance.block
+    }
 };