+++ packages/nimble-components/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @ni/nimble-components
 
-This log was last generated on Mon, 20 Jun 2022 23:17:47 GMT and should not be manually modified.
+This log was last generated on Mon, 27 Jun 2022 17:17:28 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 11.4.0
+
+Mon, 27 Jun 2022 17:17:28 GMT
+
+### Minor changes
+
+- New token for divider background color ([ni/nimble@12ed391](https://github.com/ni/nimble/commit/12ed391e04e9bdaac550d401704d54e0b6a10e83))
+
+## 11.3.0
+
+Fri, 24 Jun 2022 15:18:24 GMT
+
+### Minor changes
+
+- Implementation, styling, and tests for the 'default' state of the nimble-tooltip. ([ni/nimble@70e8d55](https://github.com/ni/nimble/commit/70e8d554250b8abc4ce7bdc069dcbf28107fee7c))
+
+## 11.2.0
+
+Thu, 23 Jun 2022 18:40:01 GMT
+
+### Minor changes
+
+- Add appearances and update inc/dec for number field ([ni/nimble@2678f74](https://github.com/ni/nimble/commit/2678f7470e090ed20520a4cd98fc07a8c2ab995b))
 
 ## 11.1.1